+++ app/src/main/kotlin/com/taskworld/android/restfulandroidkotlin/helper/message/ProductMessages.kt
@@ -1,19 +1,18 @@
 package com.taskworld.android.restfulandroidkotlin.helper.message
 
-import com.taskworld.android.restfulandroidkotlin.helper.InsertMessage
-import com.taskworld.android.restfulandroidkotlin.helper.UpdateMessage
-import com.taskworld.android.restfulandroidkotlin.helper.DeleteMessage
+import com.taskworld.android.restfulandroidkotlin.helper.message.InsertMessage
+import com.taskworld.android.restfulandroidkotlin.helper.message.UpdateMessage
+import com.taskworld.android.restfulandroidkotlin.helper.message.DeleteMessage
 import com.taskworld.android.restfulandroidkotlin.Product
 
 /**
  * Created by VerachadW on 10/24/2014 AD.
  */
 
-val TABLE_NAME = javaClass<Product>().getSimpleName()
 
-class InsertProductMessage(key: String): InsertMessage(TABLE_NAME, key)
+class InsertProductMessage(key: String): InsertMessage(Product.TABLE_NAME, key)
 
-class UpdateProductMessage(key: String, changeSet: Map<String, String>): UpdateMessage(TABLE_NAME, key, changeSet)
+class UpdateProductMessage(key: String, changeSet: Map<String, String>): UpdateMessage(Product.TABLE_NAME, key, changeSet)
 
-class DeleteProductMessage(key: String): DeleteMessage(TABLE_NAME, key)
+class DeleteProductMessage(key: String): DeleteMessage(Product.TABLE_NAME, key)
